refactor[#3]: 코드 리팩토링
