+++ backend/src/main/java/com/handong/rebon/exception/tag/TagException.java
@@ -4,7 +4,7 @@ import org.springframework.http.HttpStatus;
 import com.handong.rebon.exception.RebonException;
 
 public abstract class TagException extends RebonException {
-    
+
     protected TagException(String message, HttpStatus httpStatus) {
         super(message, httpStatus);
     }
