--- backend/src/main/java/com/handong/rebon/tag/application/dto/TagRequestDto.java
@@ -1,12 +1,0 @@
-package com.handong.rebon.tag.application.dto;
-
-import lombok.Getter;
-
-@Getter
-public class TagRequestDto {
-    private String name;
-
-    public TagRequestDto(String name){
-        this.name = name;
-    }
-}
