+++ backend/src/main/java/com/handong/rebon/tag/domain/ShopTag.java
@@ -1,12 +1,14 @@
 package com.handong.rebon.tag.domain;
 
 import javax.persistence.*;
+
 import com.handong.rebon.shop.domain.item.Shop;
 
 @Entity
 public class ShopTag {
 
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @ManyToOne
