+++ backend/src/main/java/com/handong/rebon/tag/domain/Tag.java
@@ -1,9 +1,11 @@
 package com.handong.rebon.tag.domain;
 
 import com.handong.rebon.exception.tag.TagNameException;
+
 import java.util.ArrayList;
 import java.util.List;
 import javax.persistence.*;
+
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
