+++ backend/src/main/java/com/handong/rebon/tag/domain/repository/ShopTagRepository.java
@@ -3,5 +3,5 @@ package com.handong.rebon.tag.domain.repository;
 import com.handong.rebon.tag.domain.ShopTag;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ShopTagRepository extends JpaRepository<ShopTag, Long>{
+public interface ShopTagRepository extends JpaRepository<ShopTag, Long> {
 }
