--- backend/src/main/java/com/handong/rebon/tag/presentation/AdminShopTagController.java
@@ -1,14 +1,0 @@
-package com.handong.rebon.tag.presentation;
-
-import com.handong.rebon.tag.application.ShopTagService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@RequiredArgsConstructor
-@RequestMapping("/admin/shoptags")
-@Controller
-public class AdminShopTagController {
-
-    private final ShopTagService shopTagService;
-}
