+++ backend/src/test/java/com/handong/rebon/integration/tag/TagIntegrationTest.java
@@ -8,8 +8,10 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
+
 import javax.transaction.Transactional;
 
 @SpringBootTest
@@ -33,7 +35,7 @@ public class TagIntegrationTest {
         Tag newTag = tagRepository.getById(id);
 
         // then
-        assertThat(newTag.getName()).isEqualTo(tagName);
+        assertThat(newTag).extracting("name").isEqualTo(tagName);
     }
 
     @Test
