+++ backend/src/test/java/com/handong/rebon/unit/tag/domain/TagTest.java
@@ -1,17 +1,18 @@
-package com.handong.rebon.unit.tag.domain.content;
+package com.handong.rebon.unit.tag.domain;
 
 import com.handong.rebon.exception.tag.TagNameException;
 import com.handong.rebon.tag.domain.Tag;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
-public class TagContentTest {
+public class TagTest {
 
     @ParameterizedTest
     @ValueSource(strings = {"", " "})
-    @DisplayName("태그 이름이 null이거나 빈 문자열이면 생성할 수 없다.")
+    @DisplayName("태그 이름이 빈 문자열이면 생성할 수 없다.")
     public void createBlankTag(String name) {
         //given
         //when, then
